requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.1
Pillow==10.0.1
proglog==0.1.10
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
